add standard popup when someone tries to remove a comment or post

+++ src/components/ToggleMWmenu.tsx
@@ -346,7 +346,7 @@ export const ToggleMWmenu: React.FC<ToggleMenuMWPropsType> = ({
                       </div>
                       <span
                         onClick={(e) => {
-                          // if (window.confirm('Are You Sure?'))
+                          if (window.confirm('Are You Sure?'))
                             onRemove(LoungeId, 'D');
                         }}
                       >
@@ -440,7 +440,7 @@ export const ToggleMWmenu: React.FC<ToggleMenuMWPropsType> = ({
                 <li
                   className='nav-item'
                   onClick={(e) => {
-                    // if (window.confirm('Are You Sure?'))
+                    if (window.confirm('Are You Sure?'))
                       onRemove(LoungeId, 'P');
                   }}
                 >
@@ -461,11 +461,11 @@ export const ToggleMWmenu: React.FC<ToggleMenuMWPropsType> = ({
                       </div>
                       <span
                         onClick={(e) => {
-                          // if (
-                          //   window.confirm(
-                          //     'This will remove post image Are you sure?'
-                          //   )
-                          // )
+                          if (
+                            window.confirm(
+                              'This will remove post image Are you sure?'
+                            )
+                          )
                             removePostImage(LoungeId);
                         }}
                       >
